Add a serial communication example
TX is on GPIO 1, that is J1 pin 1
RX is on GPIO 2, that is J1 pin 2

Speed set to 115200 bps

To use connect a USB BUB (or similar), RX to J1-1, TX to J1-2 and
ensure that it's configured for 3.3 V signaling.

Note, this came about after struggling and failing to get JTAG UART
working.  Always end up going back to good 'ol serial.